Add files via upload
Added Data_Mods class. This class contains the various methods for modifying the raw data (only moving all starting palm locations to a common point currently implemented). It also has the method to create vectors from the data. Vectors can be done based on frames (resulting in an array where each frame is a column so if the current raw data was first_middle_last and you want only the middle frames you'd use [:,1,:]. I have not added a method to randomize the data yet.